Working version

+++ notes.py
@@ -6,87 +6,88 @@ import subprocess
 from typing import Dict, Tuple
 
 
-class Error(Exception):
+class notes_Error(Exception):
     pass
 
 
-def encrypt_text(login: str, master_key: bytes,
+def _encrypt_text(login: str, master_key: bytes,
                  text: str) -> bytes:
     if not isinstance(login, str) or not isinstance(master_key, bytes) \
             or not isinstance(text, str):
-        raise Error("Error in notes.encrypt_text(): Invalid input type")
+        raise notes_Error("Error in notes._encrypt_text(): Invalid input type")
 
     text_bytes = text.encode(encoding="utf-8")
     if not isinstance(text_bytes, bytes):
-        raise Error("Error in notes.encrypt_text(): Incorrect text encoding")
+        raise notes_Error(
+            "Error in notes._encrypt_text(): Invalid text encoding")
 
     try:
         checker1 = db.info(login)
         if checker1 is None:
-            raise Error(f"Error in notes.encrypt_text(): \
+            raise notes_Error(f"Error in notes._encrypt_text(): \
                         User with login {login} not found")
-    except Error as e:
-        raise Error(str(e))
+    except db.db_Error as e:
+        raise notes_Error(str(e))
     info: Tuple[object, ...] = tuple(checker1)
 
     encrypted_data: Dict[str, bytes] = {"ciphertext": info[3], "iv": info[4]}
 
     try:
         checker2 = security.decrypt(encrypted_data, master_key)
-    except Error as e:
-        raise Error(str(e))
+    except security.security_Error as e:
+        raise notes_Error(str(e))
     key: bytes = checker2
 
     try:
         checker3 = security.encrypt(text_bytes, key, info[4])
-    except Error as e:
-        raise Error(str(e))
+    except security.security_Error as e:
+        raise notes_Error(str(e))
     ct: bytes = checker3["ciphertext"]
     if not isinstance(ct, bytes):
-        raise Error("Error in notes.encrypt_text(): Invalid output type")
+        raise notes_Error("Error in notes._encrypt_text(): Invalid output type")
     return ct
 
 
-def decrypt_text(login: str, master_key: bytes, ct: bytes) -> str:
+def _decrypt_text(login: str, master_key: bytes, ct: bytes) -> str:
     if not isinstance(login, str) or not isinstance(master_key, bytes) \
             or not isinstance(ct, bytes):
-        raise Error("Error in notes.decrypt_text(): Invalid input type")
+        raise notes_Error("Error in notes._decrypt_text(): Invalid input type")
 
     try:
         checker1 = db.info(login)
         if checker1 is None:
-            raise Error(f"Error in notes.encrypt_text(): \
+            raise notes_Error(f"Error in notes._encrypt_text(): \
                         User with login {login} not found")
-    except Error as e:
-        raise Error(str(e))
+    except db.db_Error as e:
+        raise notes_Error(str(e))
     info: Tuple[object, ...] = tuple(checker1)
 
     encrypted_data: Dict[str, bytes] = {"ciphertext": info[3], "iv": info[4]}
 
     try:
         checker2 = security.decrypt(encrypted_data, master_key)
-    except Error as e:
-        raise Error(str(e))
+    except security.security_Error as e:
+        raise notes_Error(str(e))
     key: bytes = checker2
 
     encrypted_text = {"ciphertext": ct, "iv": info[4]}
 
     try:
         checker3 = security.decrypt(encrypted_text, key)
-    except Error as e:
-        raise Error(str(e))
+    except security.security_Error as e:
+        raise notes_Error(str(e))
     text: bytes = checker3
 
     result = text.decode(encoding="utf-8")
 
     if not isinstance(result, str):
-        raise Error("Error in notes.decrypt_text(): Invalid output type")
+        raise notes_Error("Error in notes._decrypt_text(): Invalid output type")
     return result
 
 
-def visual(path: str) -> None:
+def _visual(path: str) -> None:
     if not isinstance(path, str):
-        raise Error("Error in notes.visual(): Invalid input type")
+        raise notes_Error("Error in notes._visual(): Invalid input type")
     path_ = path + '_temp'
     if os.name == 'posix':
         soft = ['kate', 'vim', 'geany', 'gedit', 'nano']
@@ -99,30 +100,35 @@ def visual(path: str) -> None:
     elif os.name == 'nt':
         m = subprocess.run(['notepad', path_])
         return None
-    raise Error("Error in notes.visual(): Operating system not supported")
+    raise notes_Error(
+        "Error in notes._visual(): Operating system not supported")
 
 
 def write(login: str, master_key: bytes, name: str) -> None:
     if not isinstance(login, str) or not isinstance(name, str) \
             or not isinstance(master_key, bytes):
-        raise Error("Error in notes.write(): Invalid input type")
+        raise notes_Error("Error in notes.write(): Invalid input type")
 
-    path_ = os.path.join("authentication/notes", login, name)
+    path_ = os.path.join('authentication', 'notes', login, name)
     if os.path.exists(path_):
         os.remove(path_)
 
     try:
-        visual(path_)
-    except Error as e:
-        raise Error(str(e))
-    with open(path_ + '_temp', "r", encoding="utf-8") as t:
+        _visual(path_)
+    except notes_Error as e:
+        raise notes_Error(str(e))
+    if not os.path.exists(path_ + "_temp"):
+        raise notes_Error(
+            "Error in notes.write(): Invalid work of the text editor")
+
+    with open(path_ + "_temp", 'r', encoding="utf-8") as t:
         text: str = t.read()
-    os.remove(path_ + '_temp')
+    os.remove(path_ + "_temp")
 
     try:
-        checker2 = encrypt_text(login, master_key, text)
-    except Error as e:
-        raise Error(str(e))
+        checker2 = _encrypt_text(login, master_key, text)
+    except notes_Error as e:
+        raise notes_Error(str(e))
     enc_text: bytes = checker2
 
     with open(path_, "wb") as f:
@@ -130,14 +136,14 @@ def write(login: str, master_key: bytes, name: str) -> None:
     return None
 
 
-def read(login: str, master_key: bytes, name: str) -> None:
+def edit(login: str, master_key: bytes, name: str) -> None:
     if not isinstance(login, str) or not isinstance(name, str) \
             or not isinstance(master_key, bytes):
-        raise Error("Error in notes.read(): Invalid input type")
+        raise notes_Error("Error in notes.edit(): Invalid input type")
 
-    path_ = os.path.join("authentication/notes", login, name)
+    path_ = os.path.join('authentication', 'notes', login, name)
     if not os.path.exists(path_):
-        raise Error("Error in notes.read(): Note does not exist")
+        raise notes_Error("Error in notes.edit(): Note does not exist")
 
     enc_text = bytes()
     with open(path_, "rb") as f:
@@ -149,18 +155,18 @@ def read(login: str, master_key: bytes, name: str) -> None:
             pass
 
     try:
-        checker1 = decrypt_text(login, master_key, enc_text)
-    except Error as e:
-        raise Error(str(e))
+        checker1 = _decrypt_text(login, master_key, enc_text)
+    except notes_Error as e:
+        raise notes_Error(str(e))
     text: str = str(checker1)
 
     with open(path_ + '_temp', "w", encoding="utf-8") as t:
         t.write(text)
 
     try:
-        visual(path_)
-    except Error as e:
-        raise Error(str(e))
+        _visual(path_)
+    except notes_Error as e:
+        raise notes_Error(str(e))
 
     with open(path_ + '_temp', "r", encoding="utf-8") as t:
         new_text: str = t.read()
@@ -168,9 +174,9 @@ def read(login: str, master_key: bytes, name: str) -> None:
     os.remove(path_ + '_temp')
 
     try:
-        checker3 = encrypt_text(login, master_key, new_text)
-    except Error as e:
-        raise Error(str(e))
+        checker3 = _encrypt_text(login, master_key, new_text)
+    except notes_Error as e:
+        raise notes_Error(str(e))
     new_enc_text: bytes = checker3
 
     with open(path_, "wb") as f:
@@ -178,13 +184,39 @@ def read(login: str, master_key: bytes, name: str) -> None:
     return None
 
 
+def note_list(login: str) -> Tuple[str, ...]:
+    if not isinstance(login, str):
+        raise notes_Error("Error in notes.note_list(): Invalid input type")
+
+    path_ = os.path.join('authentication', 'notes', login)
+    notes = tuple(os.listdir(path_))
+    return notes
+
+
 def delete(login: str, name: str) -> None:
     if not isinstance(login, str) or not isinstance(name, str):
-        raise Error("Error in notes.delete(): Invalid input type")
+        raise notes_Error("Error in notes.delete(): Invalid input type")
 
-    path_ = os.path.join("authentication/notes", login, name)
+    path_ = os.path.join('authentication', 'notes', login, name)
     if os.path.exists(path_):
         os.remove(path_)
         return None
     else:
-        raise Error("Error in notes.delete(): Note does not exist")
+        raise notes_Error("Error in notes.delete(): Note does not exist")
+
+
+def delete_all(login: str) -> None:
+    if not isinstance(login, str):
+        raise notes_Error("Error in notes.delete_all(): Invalid input type")
+
+    path_ = os.path.join('authentication', 'notes', login)
+    notes = tuple(os.listdir(path_))
+
+    for note in notes:
+        path_ = os.path.join('authentication', 'notes', login, note)
+        if os.path.exists(path_):
+            os.remove(path_)
+        else:
+            raise notes_Error(
+                "Error in notes.delete_all(): Note does not exist")
+    return None
